Review/requirements.txt
Flask==2.3.3
mysql-connector-python==8.0.33
python-dotenv==1.0.0  # Optional: For loading environment variables from a .env file
Flask-WTF==1.1.1  # Optional: For handling forms in a secure way
